Move TerminateInsuranceNavHost into /navigation package
